services/signal-service/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
psycopg[binary,pool]>=3.2.0
python-dotenv>=1.0.1
websockets>=14.0
